V1.1.3  修改模拟IIC默认引脚  修改ips200_displayimage8660_zoom1函数的第一个参数的类型为uint16  修复PIT1不可用的问题

+++ Example/printf使用/Libraries/seekfree_peripheral/SEEKFREE_IPS200_PARALLEL8.h
@@ -161,7 +161,7 @@ void ips200_displayimage032(uint8 *p, uint16 width, uint16 height);
 void ips200_displayimage032_zoom(uint8 *p, uint16 width, uint16 height, uint16 dis_width, uint16 dis_height);
 void ips200_displayimage032_zoom1(uint8 *p, uint16 width, uint16 height, uint16 start_x, uint16 start_y, uint16 dis_width, uint16 dis_height);
 void ips200_displayimage8660_zoom(uint16 *p, uint16 width, uint16 height, uint16 dis_width, uint16 dis_height);
-void ips200_displayimage8660_zoom1(uint8 *p, uint16 width, uint16 height, uint16 start_x, uint16 start_y, uint16 dis_width, uint16 dis_height);
+void ips200_displayimage8660_zoom1(uint16 *p, uint16 width, uint16 height, uint16 start_x, uint16 start_y, uint16 dis_width, uint16 dis_height);
 void ips200_displayimage8660_grayscale_zoom(uint16 *p, uint16 width, uint16 height, uint16 dis_width, uint16 dis_height);
 void ips200_displayimage7725(uint8 *p, uint16 width, uint16 height);
 void ips200_display_chinese(uint16 x, uint16 y, uint8 size, const uint8 *p, uint8 number, uint16 color);
